Individual Table Done
The functionality to roll the individual monster loot table implemented with s call to the table function and the type  parameter 'ind' and you may enter any monster CR to determine which table to use.

NOTES: might want to make those ugly if statements a function that can be called that would return row. May want to test the random function as well as the overhead of a high amount of calls.